requirements.txt
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==9.2.0
streamlit==1.10.0
tensorflow==2.9.1
